+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added more color grading tools
 - Added changelog level for Shader Variant stripping
 - Added an emissive color node and a slider to control the pre-exposure percentage of emission color
+- Added Debug mode for validation of material albedo and metalness/specularColor values
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering